chore: various cleanups
