+++ examples/terraformwithmodules/modules/instance/modules/childinstance/main.tf
@@ -1,0 +1,20 @@
+variable "instance_type" {
+  type = string
+}
+
+
+resource "aws_instance" "web_app" {
+  ami           = "ami-674cbc1e"
+  instance_type = var.instance_type
+
+  root_block_device {
+    volume_size = 50
+  }
+
+  ebs_block_device {
+    device_name = "my_data"
+    volume_type = "io1"
+    volume_size = 1000
+    iops        = 800
+  }
+}
